requirements.txt
discord.py==2.3.2
python-dotenv==1.1.1
aiosqlite==0.19.0
pytz==2024.1
flask==3.0.0 